+++ src/Network/Network/Models/PSNetworkInterface.cs
@@ -43,6 +43,9 @@ namespace Microsoft.Azure.Commands.Network.Models
         [Ps1Xml(Target = ViewControl.Table)]
         public bool? EnableIPForwarding { get; set; }
 
+        [Ps1Xml(Target = ViewControl.Table)]
+        public string DisableTcpStateTracking { get; set; }
+
         public List<string> HostedWorkloads { get; set; }
 
         public PSNetworkSecurityGroup NetworkSecurityGroup { get; set; }