+++ src/Network/Network/NetworkManager/NetworkManagerSecurityAdminConfiguration/NewAzNetworkManagerSecurityAdminConfigurationCommand.cs
@@ -18,6 +18,7 @@ using Microsoft.Azure.Commands.Network.Models.NetworkManager;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Commands.ResourceManager.Common.Tags;
 using Microsoft.Azure.Management.Network;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -65,8 +66,8 @@ namespace Microsoft.Azure.Commands.Network
         [Parameter(
            Mandatory = false,
            ValueFromPipelineByPropertyName = true,
-           HelpMessage = "ApplyOnNetworkIntentPolicyBasedServices. Valid values include 'None' and 'All'.")]
-        public string[] ApplyOnNetworkIntentPolicyBasedService { get; set; }
+           HelpMessage = "How security admin rules are applied to virtual networks with services using network intent policies. Valid values include 'None', 'All', and 'AllowRulesOnly'.")]
+        public NetworkIntentPolicyBasedServiceType[] ApplyOnNetworkIntentPolicyBasedService { get; set; }
 
         [Parameter(
            Mandatory = false,
@@ -81,6 +82,13 @@ namespace Microsoft.Azure.Commands.Network
 
         [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
         public SwitchParameter AsJob { get; set; }
+
+        public enum NetworkIntentPolicyBasedServiceType
+        {
+            None,
+            All,
+            AllowRulesOnly,
+        }
 
         public override void Execute()
         {
@@ -103,7 +111,16 @@ namespace Microsoft.Azure.Commands.Network
         {
             var securityConfig = new PSNetworkManagerSecurityAdminConfiguration();
             securityConfig.Name = this.Name;
-            securityConfig.ApplyOnNetworkIntentPolicyBasedServices = this.ApplyOnNetworkIntentPolicyBasedService.ToList();
+
+            if (this.ApplyOnNetworkIntentPolicyBasedService != null)
+            {
+                securityConfig.ApplyOnNetworkIntentPolicyBasedServices = new List<string>();
+                foreach (NetworkIntentPolicyBasedServiceType nipType in this.ApplyOnNetworkIntentPolicyBasedService)
+                {
+                    securityConfig.ApplyOnNetworkIntentPolicyBasedServices.Add(nipType.ToString());
+                }
+            }
+
             if (!string.IsNullOrEmpty(this.Description))
             {
                 securityConfig.Description = this.Description;