+++ src/Resources/Resources/ChangeLog.md
@@ -20,6 +20,10 @@
 
 ## Upcoming Release
 
+## Version 6.3.0
+* Removed unused reference to resource identifier in role assignment create cmdlet
+* Exposed parameter `-Count` for `Get-AzADApplication`, `Get-AzADServicePrincipal`, `Get-AzADGroup` [#19476] 
+
 ## Version 6.2.0
 * Fixed bug `-Password` overwrite `-PasswordProfile` in `New-AzADUser` [#19265]
 * Exposed `EmployeeOrgData` `Manager` for `Get-AzADUSer` [#18205]